# Conflicts:
#	public_html/admin/catalog.app/edit_product.inc.php
#	public_html/admin/currencies.app/edit_currency.inc.php
#	public_html/admin/customers.app/customers.inc.php
#	public_html/admin/customers.app/edit_customer.inc.php
#	public_html/admin/languages.app/edit_language.inc.php
#	public_html/admin/orders.app/order_statuses.inc.php
#	public_html/admin/orders.app/orders.inc.php
#	public_html/admin/slides.app/edit_slide.inc.php
#	public_html/ext/trumbowyg/langs/es_ar.min.js
#	public_html/ext/trumbowyg/langs/fr.min.js
#	public_html/ext/trumbowyg/langs/he.min.js
#	public_html/ext/trumbowyg/langs/ro.min.js
#	public_html/ext/trumbowyg/langs/rs.min.js
#	public_html/ext/trumbowyg/langs/rs_latin.min.js
#	public_html/ext/trumbowyg/langs/tr.min.js
#	public_html/ext/trumbowyg/langs/zh_cn.min.js
#	public_html/ext/trumbowyg/plugins/base64/trumbowyg.base64.min.js
#	public_html/ext/trumbowyg/plugins/colors/trumbowyg.colors.min.js
#	public_html/ext/trumbowyg/plugins/colors/ui/trumbowyg.colors.min.css
#	public_html/ext/trumbowyg/trumbowyg.min.js
#	public_html/ext/trumbowyg/ui/trumbowyg.min.css
#	public_html/includes/app_header.inc.php
#	public_html/includes/boxes/box_similar_products.inc.php
#	public_html/includes/functions/func_draw.inc.php
#	public_html/includes/library/lib_cache.inc.php
#	public_html/includes/library/lib_database.inc.php
#	public_html/includes/library/lib_document.inc.php
#	public_html/includes/templates/default.catalog/styles/application.css
#	public_html/includes/templates/default.catalog/styles/boxes.css
#	public_html/includes/templates/default.catalog/styles/checkout.css
#	public_html/includes/templates/default.catalog/styles/product.css
#	public_html/includes/templates/default.catalog/views/box_category.inc.php
#	public_html/includes/templates/default.catalog/views/box_category_tree.inc.php
#	public_html/includes/templates/default.catalog/views/box_checkout_payment.inc.php
#	public_html/includes/templates/default.catalog/views/box_checkout_shipping.inc.php
#	public_html/includes/templates/default.catalog/views/box_contact_us.inc.php
#	public_html/includes/templates/default.catalog/views/box_information.inc.php
#	public_html/includes/templates/default.catalog/views/box_manufacturer.inc.php
#	public_html/includes/templates/default.catalog/views/box_product.inc.php
#	public_html/includes/templates/default.catalog/views/box_site_footer.inc.php
#	public_html/install/data.sql
#	public_html/pages/ajax/checkout_customer.html.inc.php
#	public_html/pages/ajax/checkout_payment.html.inc.php
#	public_html/pages/ajax/checkout_shipping.html.inc.php
#	public_html/pages/category.inc.php
#	public_html/pages/checkout.inc.php
#	public_html/pages/customer_service.inc.php
#	public_html/pages/information.inc.php
#	public_html/pages/manufacturer.inc.php
#	public_html/pages/manufacturers.inc.php
#	public_html/pages/order_history.inc.php
#	public_html/pages/product.inc.php

+++ public_html/admin/translations.app/search.inc.php
@@ -3,12 +3,12 @@
   if (empty($_GET['page'])) $_GET['page'] = 1;
   if (empty($_GET['languages'])) $_GET['languages'] = array_keys(language::$languages);
   
-  if (isset($_POST['save'])) {
+  if (isset($_POST['save']) && !empty($_POST['translations'])) {
     
     foreach ($_POST['translations'] as $translation) {
       $sql_update_fields = '';
-      foreach (array_keys(language::$languages) as $language) {
-        $sql_update_fields .= "text_".$language ." = '". database::input(trim($translation['text_'.$language]), !empty($translation['html']) ? true : false) ."', ";
+      foreach ($_GET['languages'] as $language_code) {
+        $sql_update_fields .= "text_".database::input($language_code) ." = '". database::input(trim($translation['text_'.database::input($language_code)]), !empty($translation['html']) ? true : false) ."', " . PHP_EOL;
       }
       database::query(
         "update ". DB_TABLE_TRANSLATIONS ."
@@ -29,7 +29,7 @@
     exit;
   }
   
-  if (isset($_POST['delete'])) {
+  if (isset($_POST['delete']) && !empty($_POST['translation_id'])) {
     database::query(
       "delete from ". DB_TABLE_TRANSLATIONS ."
       where id = '". database::input($_POST['translation_id']) ."'
@@ -44,8 +44,15 @@
     exit;
   }
   
-  $languages = array_keys(language::$languages);
-  if (array_search('en', $languages) === false) array_unshift($languages, 'en');
+  $languages_query = database::query(
+    "select code from ". DB_TABLE_LANGUAGES ."
+    order by priority;"
+  );
+
+  $language_codes = array();
+  while ($language = database::fetch($languages_query)) {
+    $language_codes[] = $language['code'];
+  }
 ?>
 <style>
 ul.filter li {
@@ -104,7 +111,7 @@ ul.filter li {
   
   <li>
     <?php echo language::translate('title_languages', 'Languages'); ?>:<br />
-    <?php foreach ($languages as $language_code) echo '<label>'. functions::form_draw_checkbox('languages[]', $language_code) .''. $language_code .'</label>'; ?>
+    <?php foreach ($language_codes as $language_code) echo '<label>'. functions::form_draw_checkbox('languages[]', $language_code) .''. $language_code .'</label>'; ?>
   </li>
   
   <li><?php echo functions::form_draw_button('filter', language::translate('title_filter', 'Filter'), 'submit'); ?></li>
